Refactor the add deck modal into separate component
Also add types to replace 'any' in some places.
